feat: refactor end-to-end test runner script
Reason for change:
The end-to-end test runner script was long, hard to maintain and did
not share code with other scripts that "run" LDAR-Sim.

Resolution:
Refactored the end-to-end test runner script to use an end-to-end
test runner simulation manager that inherits core functionality from
the base simulation manager. This allows the end-to-end test runner
simulation manager to inherit changes made to LDAR-Sim run
functionality.

Effect(s) of change:
N/A