test: Update bad test import path
Reason for change:
The import path for a method used in unit tests was a tertiary import
(ie: importing from a file that imports from the file that contains
the method). This is bad practice and should be avoided.

Resolution:
The import path was updated to import the method directly from the file.

Effect(s) of change:
N/A