style: remove unnecessary function passing
Reason for change:
The compare_outputs function was previously being passed to the test
suite run manager due to import constraints, however these constraints
no longer apply due to moving the manager to the testing directory.

Resolution:
The compare_outputs function is no longer passed to the test suite run
manager, as it is now imported directly in the test suite run
manager.

Effect(s) of change:
N/A