feat: Add error logging functionality
Reason for change:
LDAR-Sim runtime errors can be difficult to debug when only printed to
the terminal.

Resolution:
Add  logging configuration functionality to LDAR-Sim to log errors to
a timestamped file in the logs directory and a log file in a logs
directory in the output directory.

Effect(s) of change:
Logging functionality introduced.