fix: Remove log folder from program folders
Reason for change:
Runtime error was occurring due to the logs folder in the outputs
being treated as a program folder.

Resolution:
After all folders in the outputs folder are read in, the logs folder
is removed from the list of program folders.

Effect(s) of change:
Simulation runs as expected without runtime errors.