test: Update tests failing to switch to logging
Reason for change:
Some unit tests testing for a specific error message being printed
before a system exit were failing. This was due to the fact that the
message is now being logged instead of printed.

Resolution:
Updated the failing tests to configure the log to log to the error
stream and then test if the error stream contains the expected message.

Effect(s) of change:
All tests now pass.