fix: Add option to toggle logging in output dir
Reason for change:
Not all LDAR-Sim run script managers want to log errors in the outputs.

Resolution:
Added a boolean flag to the simulation manager to toggle error logging.

Effect(s) of change:
The simulation manager will now log errors to the output directory
if the flag is set to True.