feat: Change rolling average startup logic
Reason for change:
Rolling average startup logic was working in such a way that anomalous
behavior was observed in the start of the simulation.

Resolution:
Changed rolling average startup logic to require the full
rolling average window to be filled before the rolling average
is calculated. This will prevent the rolling average from being
calculated with fewer than the full window of data.

Effect(s) of change:
The rolling average startup logic will now require the full
rolling average window to be filled before the rolling average
is calculated.