feat: Add parameter to disable pre-sim emissions
Reason for change:
In some cases, it is incorrect to simulate emissions beginning from
a state of no LDAR.

Resolution:
Added a parameter to the virtual world that allows the user to
specify whether or not to simulate emissions that "begin" before the
start of the simulation.

Effect(s) of change:
The user can now specify whether or not to simulate emissions that
"begin" before the start of the simulation.