fix: Fix runtime error if no long threshold
Reason for change:
Runtime error was occurring when the
stationary long threshold was not set.

Resolution:
Added a check to see if the stationary long threshold is defined before
using it in the rolling average calculation.

Effect(s) of change:
The program will no longer crash if the stationary long threshold
is not set.