fix: Add shape and empty df checks in date calcs
Reason for change:
When calculating start dates for estimated emissions, runtime errors
were being thrown when the shape of the dataframe was not as expected.

Resolution:
This change adds checks for the shape of the dataframe and for empty
dataframes to prevent these errors.

Effect(s) of change:
Prevents runtime errors.