fix: fixed issue with METEC_NoWind_sensor.py
Reason for change:
The METEC_NoWind_sensor was out of date for the current version of the
LDAR-Sim, resulting in a runtime error when using it at the equipment
or component level.

Resolution:
The METEC_NoWind_sensor was updated to the current version of the
LDAR-Sim, allowing it to be used at the equipment or component level.

Effect(s) of change:
The METEC_NoWind_sensor can now be used at the equipment or component
level without causing a runtime error.